Crate sctp

This crate provides high level SCTP networking. Currently it only supports basic SCTP features like multi-homing in one-to-one and one-to-many associations. SCTP notifications and working directly on associations is not supported yet but is in the TODO list.

Structs

Incoming
Iterator over incoming connections on SctpListener

SctpEndpoint
One-to-many SCTP endpoint.

SctpListener
SCTP listener which behaves like a TcpListener. A SCTP listener is used to wait for and accept one-to-one SCTP connections. An accepted connection is represented by SctpStream.

SctpStream
One-to-one SCTP connected stream which behaves like a TCP stream. A SctpStream can be obtained either actively by connecting to a SCTP endpoint with the connect constructor, or passively from a SctpListener which accepts new connections

Enums

SoDirection
Socket direction